refactor testhttp package to make it more user frendly
